+++ CHANGELOG.md
@@ -9,6 +9,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
+- Added `DAPP_ADDRESS`, `DAPP_DEPLOY_BLOCK_HASH`, `HISTORY_ADDRESS`, `AUTHORITY_ADDRESS`, and
+  `INPUT_BOX_ADDRESS` env vars to the dispatcher and the authority claimer.
+  These are optional and will overwrite the values from the dapp and rollups deployment files.
 - Added `cartesi-rollups-node` Go binary as a single entrypoint to execute all Cartesi Node services
 - Added `authority-claimer` service
 - Added `CHAIN_ID` environment variable to dispatcher config